some other fixes

+++ src/main/java/com/swanimobiliaria/model/service/PropertyService.java
@@ -8,25 +8,32 @@ import com.swanimobiliaria.model.repository.PropertyJpaRepository;
 import com.swanimobiliaria.model.strategy.SimplePropertyStrategy;
 import com.swanimobiliaria.model.type.PropertyType;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
+import java.util.UUID;
 import java.util.stream.Collectors;
+
+import static java.lang.Math.random;
 
 @Service
 public class PropertyService {
 
-    private PropertyJpaRepository propertyJpaRepository;
-    private SimplePropertyStrategy simplePropertyStrategy;
+    private final PropertyJpaRepository propertyJpaRepository;
+    private final UserService userService;
+    private final SimplePropertyStrategy simplePropertyStrategy;
 
     @Autowired
-    public PropertyService(PropertyJpaRepository propertyJpaRepository, SimplePropertyStrategy simplePropertyStrategy) {
+    public PropertyService(PropertyJpaRepository propertyJpaRepository, UserService userService, SimplePropertyStrategy simplePropertyStrategy) {
         this.propertyJpaRepository = propertyJpaRepository;
+        this.userService = userService;
         this.simplePropertyStrategy = simplePropertyStrategy;
     }
 
-    public List<PropertyDTO> getImoveis() {
-        return propertyJpaRepository.findAll()
+    public List<PropertyDTO> getImoveis(int page, int size) {
+        return propertyJpaRepository.findAll(PageRequest.of(page, size))
                 .stream()
                 .map(PropertyConverter::fromDomainToDTO)
                 .collect(Collectors.toList());
@@ -37,12 +44,16 @@ public class PropertyService {
         return PropertyConverter.fromDomainToDTO(savedProperty);
     }
 
-    public void deleteProperty(Integer propertyId) {
-        Property property = getProperty(propertyId);
-        propertyJpaRepository.delete(property);
+    public void deleteProperty(UUID userId, Integer propertyId) {
+        UUID userById = userService.findByUserId(userId);
+        if(Objects.nonNull(userById)) {
+            Property property = getProperty(propertyId);
+            propertyJpaRepository.delete(property);
+        }
     }
 
-    public PropertyDTO updateProperty(Integer propertyId, PropertyDTO propertyDTO) {
+    public PropertyDTO updateProperty(UUID userId, Integer propertyId, PropertyDTO propertyDTO) {
+        userService.findByUserId(userId);
         Property property = getProperty(propertyId);
         Property updatedProperty = propertyJpaRepository.save(PropertyConverter.fromDTOtoDomain(property, propertyDTO));
         return PropertyConverter.fromDomainToDTO(updatedProperty);
@@ -50,10 +61,22 @@ public class PropertyService {
 
     public PropertyDTO getPropertyById(Integer propertyId) {
         Property property = getProperty(propertyId);
+        try {
+            Thread.sleep(800);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
         return PropertyConverter.fromDomainToDTO(property);
     }
 
-    public List<PropertyDTO> getPropertyByOptions(PropertyType propertyType, String city, Integer rooms, Double priceFrom, Double priceTo) {
+    public List<PropertyDTO> getPropertyByOptions(String searchParam, PropertyType propertyType, String city, Integer rooms, Double priceFrom, Double priceTo) {
+        if(Objects.nonNull(searchParam)){
+            return propertyJpaRepository
+                    .findAllBySearchTerm(searchParam)
+                    .stream()
+                    .map(PropertyConverter::fromDomainToDTO)
+                    .collect(Collectors.toList());
+        }
         PropertyDTO propertyDTO = new PropertyDTO();
         propertyDTO.setPropertyType(propertyType);
         propertyDTO.setCidade(city);
@@ -64,5 +87,11 @@ public class PropertyService {
 
     private Property getProperty(Integer propertyId) {
         return propertyJpaRepository.findById(propertyId).orElseThrow(() -> new ResourceNotFoundException("Property not found!"));
+    }
+
+    public List<PropertyDTO> getRandomProperty(){
+        long total = propertyJpaRepository.count();
+        int id = (int) ((random()/3) * total);
+        return getImoveis(id, 3);
     }
 }
